final code/ablation notebook and results/ablationfirst/requirements.txt
streamlit>=1.35.0
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
Pillow>=10.0.0
numpy>=1.24.0
torchmetrics>=1.0.0
google-generativeai>=0.5.0
onnx>=1.14.0
onnxruntime>=1.16.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
scikit-image>=0.21.0
scikit-learn>=1.3.0
